# Start up optimizations grab bag