#1016 

A bunch of startup optimizations, in my testing this brings down the startup time on a wallet with a lot of channels and a configured fedimint from 12 seconds to 6 seconds.

- 1st is mainly just adding the logs everywhere I have been using. Does an optimization for new wallets, where it'll now create the first node in the background instead of waiting for that to be initialized before returning
- 2nd backgrounds the saving of our node storage on startup, this was only done so we save any changes made to our LSP config while starting up (just migrations), these aren't the end of the world if they are missed, they'll persisted on the next startup 
- 3rd makes it so we don't block on retrying spendable outputs, these can happen async and there isn't a reason to wait
- 4th fixes two issues with fedimint startup. We would block on setting the preferred guardian, this is now done in the background. We also would download the federation info everytime but we only needed it if our db didn't have it (we were already only using the federation info if we didn't have it in our db), this fixes to only do the download if its not in our db. This also does some minor touch ups like initializing federation_id instead of doing federation_info.federation_id() since this would do a hash everytime.
- 5th changes to immediately start fetching the JWT for auth, before we would only fetch it once we got our first VSS request, this saves us about ~200ms because otherwise we wait until we have read all of indexed db until we make this call. Needed to modify the auth client to only allow one call to retrieve_new_jwt at a time, otherwise this would sometimes not complete by the time the first VSS call came in and we'd request multiple JWTs. Otherwise just adds some timing logs and tiny optimizations to indexed db
- 6th changes to sync scorer in the background rather than upfront, this made the startup ~500ms faster for me. This will still read the scorer on start up so we won't be without one until it is synced.