@TonyGiorgio 2024-02-09
<img width="465" alt="Screenshot 2024-02-09 at 4 21 02 PM" src="https://github.com/MutinyWallet/mutiny-node/assets/101225832/dc283bbc-be8f-486f-b36e-a5c646f829ed">

Just saw a stream of thousands of duplicate requests getting block height and tip until it panic'd, right after a refresh. Browser completely crashed and couldn't refresh.