@benthecarman 2024-02-10
> My on chain balance went away when I refreshed after swapping into a channel. I refreshed again, waited ~30s and then it appeared again. I think chain data needs to be cached or something? Not sure why it went away, never saw this issue before.

This happened to me once when testing the DLC stuff, I think its a bug in bdk with spending unconfirmed utxos. I assume/hope when we update it'll fix