@TonyGiorgio 2024-02-10
> > Just saw a stream of thousands of duplicate requests getting block height and tip until it panic'd, right after a refresh. Browser completely crashed and couldn't refresh.
> 
> do you have any steps to reproduce? syncing logic hasn't changed at all, just starting after we start the nostr stuff instead of before

I just did a lot of refreshes as I went through E2E testing.