@benthecarman 2024-02-10
Seems like that is happening during BDK sync, maybe just move it back to where it was until #1002  is merged